refactor: make ruff accept init unused imports
