+++ ipyvuetify/_VuetifyTemplate.py
