+++ ipyvuetify/__init__.py
@@ -1,8 +1,8 @@
-from ._version import __version__  # noqa: F401
-from .generated import *  # noqa: F401, F403
-from .Html import Html  # noqa: F401
-from .Themes import theme  # noqa: F401
-from .VuetifyTemplate import VuetifyTemplate  # noqa: F401
+from ._generated import *  # noqa: F403
+from ._Html import Html as Html
+from ._version import __version__ as __version__
+from ._VuetifyTemplate import VuetifyTemplate as VuetifyTemplate
+from .Themes import theme as theme
 
 
 def _jupyter_labextension_paths():
