+++ ipyvuetify/extra/__init__.py
@@ -1,1 +1,1 @@
-from .file_input import FileInput  # noqa: F401
+from .file_input import FileInput as FileInput
